Restructure and simplify backend architecture
