--- apps/server/src/middleware/limiters.js
@@ -1,7 +1,0 @@
-import { createRateLimiter } from "./rateLimiter.js";
-
-export const loginLimiter = createRateLimiter({
-  windowMs: 15 * 60 * 1000,
-  max: 10,
-  message: "Too many login attempts. Please try again after 15 minutes.",
-});
