+++ apps/server/src/modules/auth/middleware/credentials.js
@@ -1,4 +1,4 @@
-import allowedOrigins from "../config/allowedOrigins.js";
+import allowedOrigins from "../../../config/allowedOrigins.js";
 
 const credentials = (req, res, next) => {
     const origin = req.headers.origin;
