+++ apps/server/src/modules/auth/middleware/verifyJWT.js
