+++ apps/server/src/modules/auth/user.model.js
