+++ apps/server/src/modules/image/image.model.js
